requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.24.3
openai==1.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
ezdxf==1.1.3
PyPDF2==3.0.1
ifcopenshell==0.6.0.221122
pypdf==3.17.1
Pillow==10.0.1
openai==1.3.5
sentry-sdk==1.38.0
requests==2.31.0
typing-extensions==4.8.0
OCP==7.7.0
opencascade-python==7.7.0

# Security
PyJWT==2.8.0
cryptography==42.0.5
redis==5.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
secure==0.3.0
twilio==8.15.0
phonenumbers==8.13.31  # For phone number validation

# AWS SDK for remote backups
boto3==1.34.69

# Deployment Testing
docker==7.0.0
websockets==12.0
aiohttp==3.9.3

# Performance Testing
websockets==12.0
aiohttp==3.9.3
psutil==5.9.8
numpy==1.26.4
pandas==2.2.1
dgl==2.0.0
torch==2.2.1
tqdm==4.66.2

# Focused Validation Tests
pytest==7.4.4 
# Focused Validation Tests
pytest==7.4.4

# Existing dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
websockets>=12.0
strawberry-graphql>=0.215.0
python-socketio>=5.10.0
aiofiles>=23.2.1
pydantic>=2.5.0
jinja2>=3.1.2
python-multipart>=0.0.6
redis>=5.0.0
requests>=2.31.0
asyncio-mqtt>=0.16.1

# Core system dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
networkx>=3.1
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Database and storage
psycopg2-binary>=2.9.6
sqlalchemy>=2.0.0
alembic>=1.11.0

# Testing and development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0

# Machine learning (optional)
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0
